Authorization is completed

+++ mycontacts-backend/controllers/contactController.js
@@ -41,6 +41,10 @@ const getContactById = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if (contact.userId.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to access this contact");
+  }
   res.status(200).json(contact);
 });
 
@@ -53,6 +57,10 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  if (contact.userId.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to update this contact");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -70,6 +78,10 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  if (contact.userId.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to delete this contact");
   }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
